Edit switch command to replace .svn.
svn switch is buggy when it comes to externals. To work around this we
completely delete the .svn folder.

As a future performance improvement we could keep a .svn around for each
git-svn branch. Then when we switch we can just swap a symlink to the
correct .svn folder and run svn update.