Only detect heads on check for ggit-config branch
When cloning from a local ggit cloned repository, there was a bug
which would cause the clone to succeed, but then later the ggit
configure would fail. Now we fail before the clone.

When cloning a repository only heads are automatically tracked.
git ls-remote shows both heads and remotes, so it did not properly
detect what refs would be available to a cloned respitory.